requirement.txt
flask
ultralytics
torch
torchvision
Pillow
